reactor(search): refactor search placeholder
