+++ snippets/src/pages/search-placeholder/libs.ts
@@ -1,0 +1,61 @@
+export async function waitUntilHtmlElementExists(selector: string, timeout = 5000) {
+  const start = Date.now();
+  while (!document.querySelector(selector)?.innerHTML) {
+    if (Date.now() - start > timeout) {
+      throw new Error(`Element with selector ${selector} not found or not loaded`);
+    }
+    await delay(100);
+  }
+}
+
+
+/**
+ * Dynamic Script Injection,
+ * This may save network bandwidth and improve performance
+ */
+export function injectScript(src: string, type = 'module') {
+  // To prevent duplicate script injection
+  if (!document.querySelector(`script[src="${src}"]`)) {
+    document.body.appendChild(
+      Object.assign(document.createElement('script'), { type, src })
+    );
+  }
+}
+
+export function loadPlaceholder(selector: string) {
+  const placeholderRoot = document.querySelector<HTMLDivElement>(selector)!;
+  if (placeholderRoot.innerHTML) {
+    console.log('SearchPlaceholder: Placeholder already loaded');
+    return;
+  }
+  placeholderRoot.innerHTML = `
+  <div id="placeholder-search-box-wrapper">
+    <input
+      id="placeholder-search-box"
+      type="text"
+      class="search-box"
+      placeholder="Search post..."
+    />
+    <div class="search-modal-loading">Loading Search Modal Component... </div>
+    </div>
+`}
+
+export const delay = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
+
+export class ElementController {
+  private readonly element: HTMLDivElement;
+
+  constructor(selector: string) {
+    this.element = document.querySelector<HTMLDivElement>(selector)!;
+  }
+
+  show() {
+    // Add class 'show' to the element
+    this.element.classList.add('show');
+  }
+
+  hide() {
+    // Remove class 'show' from the element
+    this.element.classList.remove('show');
+  }
+}
