+++ snippets/src/pages/search-placeholder/main.ts
@@ -1,28 +1,21 @@
 import { searchModalEvent } from "../search-modal-event";
+import { ElementController, loadPlaceholder, injectScript, waitUntilHtmlElementExists } from "./libs";
 import './search-placeholder.css';
 
-function loadPlaceholder() {
-  const placeholderRoot = document.querySelector<HTMLDivElement>('#search-placeholder-root')!;
-  if (placeholderRoot.innerHTML) {
-    console.log('SearchPlaceholder: Placeholder already loaded');
-    return;
-  }
-  placeholderRoot.innerHTML = `
-  <div id="placeholder-search-box-wrapper">
-    <input
-      id="placeholder-search-box"
-      type="text"
-      class="search-box"
-      placeholder="Search post..."
-    />
-    <div class="search-modal-loading">Loading Search Modal Component... </div>
-    </div>
-`}
-
-const delay = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
-
+/**
+ * SearchPlaceholder, responsible for emitting commands the the shell,
+ * 
+ * Commander can be any javascript function not necessarily a react component,
+ * It just simply emits event to the shell
+ * 
+ * However, SearchPlaceholder is a JS that emits event to the shell
+ * and provide the search text field as the dummy search box for user can click on it
+ * After the user click on the dummy search box, the search modal will be shown
+ * 
+ * @returns 
+ */
 function initSearchPlaceholder() {
-  loadPlaceholder();
+  loadPlaceholder('#search-placeholder-root');
   const searchBox = document.querySelector<HTMLInputElement>('#placeholder-search-box');
   if (!searchBox) {
     console.error('SearchPlaceholder: Search box not found');
@@ -32,7 +25,8 @@ function initSearchPlaceholder() {
     searchBox.blur();
     const action = 'open';
     console.log(`SeachPlaceholder: ${action} search modal`);
-    loadSearchModal();
+    // Load React Search Modal Component
+    injectScript('/js/search.js');
     const loading = new ElementController('.search-modal-loading');
     loading.show();
     await waitUntilHtmlElementExists('#search-root');
@@ -40,43 +34,5 @@ function initSearchPlaceholder() {
     searchModalEvent.dispatch({ action });
   });
 }
-
-class ElementController {
-  private readonly element: HTMLDivElement;
-
-  constructor(selector: string) {
-    this.element = document.querySelector<HTMLDivElement>(selector)!;
-  }
-
-  show() {
-    // Add class 'show' to the element
-    this.element.classList.add('show');
-  }
-
-  hide() {
-    // Remove class 'show' from the element
-    this.element.classList.remove('show');
-  }
-}
-
-export async function waitUntilHtmlElementExists(selector: string, timeout = 5000) {
-  const start = Date.now();
-  while (!document.querySelector(selector)?.innerHTML) {
-    if (Date.now() - start > timeout) {
-      throw new Error(`Element with selector ${selector} not found or not loaded`);
-    }
-    await delay(100);
-  }
-}
-
-function loadSearchModal() {
-  // To prevent duplicate script injection
-  if (!document.querySelector('script[src="/js/search.js"]')) {
-    document.body.appendChild(
-      Object.assign(document.createElement('script'), { type: 'module', src: '/js/search.js' })
-    );
-  }
-}
-
 
 initSearchPlaceholder();
